Documentation and moved methods to EnumerableExtensions class
