--- src/With/Collections/FlattenExtension.cs
@@ -1,64 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-
-namespace With.Collections
-{
-    public static class FlattenExtension
-    {
-        /// <summary>
-        /// Returns a new array that is a one-dimensional flattening of self (recursively).
-        ///
-        ///That is, for every element that is an array, extract its elements into the new array.
-        /// </summary>
-        public static IEnumerable<T> Flatten<T>(this IEnumerable self)
-        {
-            foreach (var variable in self)
-            {
-                if (variable is T)
-                {
-                    yield return (T)variable;
-                }
-                else
-                {
-                    foreach (var result in Flatten<T>((IEnumerable)variable))
-                    {
-                        yield return result;
-                    }
-                }
-            }
-        }
-
-        /// <summary>
-        /// Returns a new array that is a one-dimensional flattening of self (recursively).
-        ///
-        ///That is, for every element that is an array, extract its elements into the new array.
-        ///
-        ///The optional level argument determines the level of recursion to flatten.
-        /// </summary>
-        public static IEnumerable Flatten(this IEnumerable self, int? order = null)
-        {
-            if (order == null || order >= 0)
-            {
-                foreach (var variable in self)
-                {
-                    if (variable is IEnumerable && !(variable is string))
-                    {
-                        foreach (var result in Flatten((IEnumerable)variable, order != null ? order - 1 : null))
-                        {
-                            yield return result;
-                        }
-                    }
-                    else
-                    {
-                        yield return variable;
-                    }
-                }
-            }
-            else
-            {
-                yield return self;
-            }
-        }
-    }
-}
-
